task_9378 环境管理光照度接口开发

+++ src/main/java/com/hoolink/sdk/bo/device/monitor/EnvironmentHistoryBO.java
@@ -99,5 +99,5 @@ public class EnvironmentHistoryBO implements Serializable {
     private Float pmTenBeyond;
 
     /** 光照度 */
-    private Integer light;
+    private Long light;
 }
